# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Qwibi x402 API",
    "version": "1.0.0",
    "description": "Pay-per-request AI analyst desk for agentic finance. Qwibi uses x402 on Algorand through the GoPlausible facilitator."
  },
  "servers": [
    {
      "url": "https://qwibi.xyz"
    }
  ],
...
  "paths": {
...
    "/api/x402/verify": {
      "get": {
        "tags": ["x402"],
        "summary": "Verify Qwibi deliverable",
        "description": "Paid x402 endpoint. Verifies a Qwibi deliverable by job id or SHA-256 attestation hash.",
        "parameters": [
          { "name": "jobId", "in": "query", "schema": { "type": "string" } },
          { "name": "hash", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "200": { "$ref": "#/components/responses/Verified" }
        }
      },
      "post": {
        "tags": ["x402"],
        "summary": "Verify Qwibi deliverable",
        "description": "Paid x402 endpoint. Verifies a Qwibi deliverable by job id or SHA-256 attestation hash.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "jobId": { "type": "string" },
                  "hash": { "type": "string" }
                },
                "anyOf": [
                  { "required": ["jobId"] },
                  { "required": ["hash"] }
                ]
              }
            }
          }
        },
        "responses": {
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "200": { "$ref": "#/components/responses/Verified" }
        }
      }
    }
  },
  "components": {
...
    "responses": {
      "PaymentRequired": {
        "description": "x402 payment required challenge"
      },
...
      "Verified": {
        "description": "Qwibi deliverable verification result",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "ok": { "type": "boolean" },
                "status": { "type": "string" },
                "verified": { "type": "boolean" },
                "untampered": { "type": "boolean" },
                "jobId": { "type": "string" },
                "attestationHash": { "type": "string" },
                "manifest": { "type": "object" },
                "verifyUrl": { "type": "string" }
              }
            }
          }
        }
      }
    }
  }
}